PDCL-7517 - move createAddIdentity from request/contentModifiers/ to request/
