+++ src/components/Identity/getIdentity/createIdentityRequestPayload.js
@@ -10,8 +10,10 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { createRequestPayload } from "../../../utils/request";
-import { createAddIdentity } from "../../../utils/request/contentModifiers";
+import {
+  createAddIdentity,
+  createRequestPayload
+} from "../../../utils/request";
 
 export default namespaces => {
   const content = {
