+++ src/utils/request/index.js
@@ -10,6 +10,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
+export { default as createAddIdentity } from "./createAddIdentity";
 export { default as createDataCollectionRequest } from "./createDataCollectionRequest";
 export { default as createDataCollectionRequestPayload } from "./createDataCollectionRequestPayload";
 export { default as createRequest } from "./createRequest";
