+++ test/unit/specs/utils/request/createAddIdentity.spec.js
@@ -1,4 +1,4 @@
-import { createAddIdentity } from "../../../../../../src/utils/request/contentModifiers";
+import { createAddIdentity } from "../../../../../src/utils/request";
 
 describe("createAddIdentity", () => {
   it("should return a function to add identity", () => {
